break out utility methods from reducer
