+++ src/reducer.js
@@ -1,4 +1,5 @@
 import {standard} from '../data/rankings.json';
+import {getListName, getCssClass, getPickRound} from './utils';
 
 export const init = {
     pick: 1,
@@ -15,43 +16,6 @@ export const init = {
     taken: []
 };
 
-export function getCssClass (vsADP) {
-    let cssClass = '';
-
-    if (vsADP.includes('+')) {
-        cssClass = 'steal';
-    }
-    else if (vsADP.includes('-')) {
-        cssClass = 'reach';
-    }
-
-    return cssClass;
-}
-
-export function getListName (action) {
-    let listName = '';
-
-    if (action === 'PLAYER_DRAFTED') {
-        listName = 'drafted';
-    }
-    else if (action === 'PLAYER_TAKEN') {
-        listName = 'taken';
-    }
-
-    return listName;
-}
-
-export function getPickRound (pick, round) {
-    if (pick >= 12) {
-        pick = 1;
-        round++;
-    } else {
-        pick++;
-    }
-
-    return [pick, round];
-}
-
 export default function reducer (state = init, action, args) {
     switch (action) {
         case 'PLAYER_DRAFTED':
@@ -64,7 +28,7 @@ export default function reducer (state = init, action, args) {
             picked.pick = `${round}-${pick} (${overall})`;
 
             state[listName].unshift(picked);
-    
+
             [pick, round] = getPickRound(pick, round);
             overall++;
